+++ src/EditeurBundle/Controller/LaboratoireController.php
@@ -56,8 +56,45 @@ class LaboratoireController extends Controller
 
         if ($editForm->isSubmitted() && $editForm->isValid()) {
 
-            $laboratoire = $editForm->getData();
-            $em = $this->getDoctrine()->getManager();
+            // --------------------------
+            //Ajout des établissements
+            // --------------------------
+
+            //s'il n'y en a qu'un, ajout automatique
+            if(count($etablissements) == 1){
+
+                $repository = $this->getDoctrine()->getRepository('AppBundle:Etablissement');
+                $etablissement = $repository->findOneByEtablissementId($etablissements[0]['id']);
+
+                $laboratoire->addEtablissement($etablissement);
+            }
+            //sinon l'utilisateur choisit lui-même un établissement
+
+            // --------------------------
+            //Set année de la collecte
+            // --------------------------
+            //vérification qu'il y a une collecte active
+            $query = $em->createQuery(
+                'SELECT COUNT(c.collecteId) as nb FROM AppBundle:Collecte c WHERE c.active = true'
+            );
+            $checkCollecte = $query->getResult();
+            $checkCollecte = $checkCollecte[0]['nb'];
+
+            //s'il y a plus que 0 = il y a une collecte active donc je prends sa date
+            if($checkCollecte > 0){
+                $query = $em->createQuery(
+                    'SELECT c.annee as annee FROM AppBundle:Collecte c WHERE c.active = true'
+                );
+            }
+            else{
+                $query = $em->createQuery(
+                    'SELECT c.annee as annee FROM AppBundle:Collecte c WHERE c.complete = true ORDER BY c.annee DESC'
+                );
+                $query->setMaxResults(1);
+            }
+            $year = $query->getSingleResult();
+            $year = $year['annee'];
+            $laboratoire->setAnneeCollecte($year);
 
             $now = new \DateTime();
             $laboratoire->setDateCreation($now);
@@ -66,13 +103,20 @@ class LaboratoireController extends Controller
             $em->persist($laboratoire);
             $em->flush();
 
+            //Création et set de l'objetId
+            $lastId = $laboratoire->getLaboId();
+            $laboratoire->setObjetId("L".$lastId);
+
+            $em->persist($laboratoire);
+            $em->flush();
 
             return $this->redirectToRoute('editeur');
         }
 
         return $this->render('EditeurBundle:Labo:new.html.twig', array(
             'edit_form' => $editForm->createView(),
-            'labo' => $laboratoire
+            'labo' => $laboratoire,
+            'etablissements' => $etablissements
         ));
     }
 
@@ -129,6 +173,7 @@ class LaboratoireController extends Controller
             'labo' => $laboratoire,
             'edit_form' => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
+            'etablissements' => $etablissements
         ));
     }
 
@@ -143,12 +188,17 @@ class LaboratoireController extends Controller
     {
 
         $em = $this->getDoctrine()->getManager();
+        $user = $this->getUser();
 
-        /** @var Labo $labo */
-        $labo = $em->getRepository('AppBundle:Labo')
-            ->find($laboId);
-        $em->remove($labo);
-        $em->flush();
+        if ($user->hasRole('ROLE_USER')){
+
+            /** @var Labo $labo */
+            $labo = $em->getRepository('AppBundle:Labo')
+                ->find($laboId);
+            $em->remove($labo);
+            $em->flush();
+        }
+
 
         return new Response(null, 204);
